pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "geo-simulations"
dynamic = ["version"]
description = "Modelling scenarios with geographic data."
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
keywords = []
authors = [
  { name = "Dirk Sidney Jansen", email = "sidisinsane@users.noreply.github.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "numpy>=1.26.4",
  "pandas>=2.2.2",
  "scipy>=1.13.1",
  "moviepy>=1.0.3",
  "requests>=2.32.2",
  "pillow==9.5.0",
  "imageio>=2.34.1",
  "rasterio>=1.3.10",
  "python-dotenv>=1.0.1",
]

[project.urls]
Documentation = "https://github.com/sidisinsane/geo-simulations#readme"
Issues = "https://github.com/sidisinsane/geo-simulations/issues"
Source = "https://github.com/sidisinsane/geo-simulations"

[tool.hatch.version]
path = "src/geo_simulations/__about__.py"

[tool.hatch.envs.default]
dependencies = [
  "coverage[toml]>=6.5",
  "pytest",
  "boto3>=1.34.113",
]

[tool.hatch.envs.default.scripts]
generate = "python src/geo_simulations/generate.py"
setup = "python src/geo_simulations/setup.py"
logger = "python src/geo_simulations/logger.py"
zombie-outbreak = "python src/geo_simulations/zombie_outbreak.py"
geotiff-metadata = "python src/geo_simulations/geotiff_metadata.py"
geotiff-convert = "python src/geo_simulations/geotiff_convert.py"
wip = "python src/geo_simulations/wip.py"
demo = "python src/geo_simulations/demo.py"
check = [
  "hatch fmt",
  "hatch run types:check",
  "hatch run security:check",
]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = [
  "- coverage combine",
  "coverage report",
]
cov = [
  "test-cov",
  "cov-report",
]

[[tool.hatch.envs.all.matrix]]
python = ["3.9", "3.10", "3.11", "3.12"]

[tool.hatch.envs.security]
dependencies = [
  "bandit>=1.7.8",
]
[tool.hatch.envs.security.scripts]
check = "bandit -c bandit.yaml -r ."


[tool.hatch.envs.types]
dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/geo_simulations tests}"

[tool.coverage.run]
source_pkgs = ["geo_simulations", "tests"]
branch = true
parallel = true
omit = [
  "src/geo_simulations/__about__.py",
]

[tool.coverage.paths]
geo_simulations = ["src/geo_simulations", "*/geo-simulations/src/geo_simulations"]
tests = ["tests", "*/geo-simulations/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
